requirements.txt
fire>=0.4.0
python-frontmatter>=1.0.0
python-slugify>=6.1.2